set up project backend to deploy
