+++ vercel.json
@@ -1,0 +1,5 @@
+{
+    "version": 2,
+    "builds": [{ "src": "api/index.js", "use": "@vercel/node" }],
+    "routes": [{ "src": "/(.*)", "dest": "/api/index.js" }]
+}
